Gemini_Code_Generator/requirements.txt
google-generativeai
PyQt5
GitPython
